+++ src/chronicle/services/__init__.py
@@ -1,4 +1,5 @@
 from chronicle.services.artifact_service import ArtifactService
+from chronicle.services.boundary_service import BoundaryService
 from chronicle.services.chronicle_service import ChronicleService
 from chronicle.services.context_service import ContextService
 from chronicle.services.decision_service import DecisionService
@@ -7,6 +8,7 @@ from chronicle.services.search_service import SearchService
 
 __all__ = [
     "ArtifactService",
+    "BoundaryService",
     "ChronicleService",
     "ContextService",
     "DecisionService",